enhances the fix introduced in 42bcff2 with gpr
By augmenting the grade_plugin_return object with fields
for the first and last filter initials, we take advantage of
existing gradebook functionality for maintaining state, esp
with forms. While the previous commit fixes one case, initial filtering
was lost when 'turning editing on or off' and when using quick edit.
The changes introduced in this commit could/should be further
applied to column-wise quick edit forms.